IndexedDB/submissions/Microsoft/idbdatabase_deleteObjectStore3.htm

<!DOCTYPE html>
<html>
<head>
    <title>IDBDatabase.deleteObjectStore() - attempt to remove an object store that does not exist </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var db = null;
        var t = async_test(); 

        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName);
            rqOpen.onupgradeneeded = t.step_func( function (event) {
                db = event.target.result;
                db.onerror = t.step_func( assert_database_error );
                try {
                    db.deleteObjectStore(objectStoreName);
                } catch (ex) {
                    assert_equals(ex.name, "NotFoundError", "ex.name");
                    t.done()
                    return;
                }
                assert_expected_exception();
            });
            rqOpen.onerror = t.step_func( function(event) { assert_open_request_error(event); } );
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
